Balanced Brackets
You will receive n lines. On those lines, you will receive one of the following:
•	Opening bracket – “(“,
•	Closing bracket – “)” or
•	Random string
Your task is to find out if the brackets are balanced. That means after every closing bracket should follow an opening one. Nested parentheses are not valid, and if two consecutive opening brackets exist, the expression should be marked as unbalanced.