Songs
Define a class Song, which holds the following information about songs: Type List, Name and Time.
On the first line you will receive the number of songs - N. 
On the next N-lines you will be receiving data in the following format: "{typeList}_{name}_{time}".
On the last line you will receive Type List / "all". Print only the Names of the songs which are from that Type List / All songs.