Battle Manager
Create a program that manages battles. You need to keep information about people, the health and energy they have. You will be receiving lines with commands until you receive the "Results" command. There are three possible commands: 
•	"Add:{personName}:{health}:{energy}":
o	Add the person, his/her health and energy to your records. If person with the given name already exists, just increase the health of the person with the current one that is given.
•	"Attack:{attackerName}:{defenderName}:{damage}":
o	Check if both people exist and if they do, reduce the defender’s health with the damage given. If the defender’s health reaches 0 or less, the person is disqualified, and you need to remove him/her from your records and print the following message:
	"{defenderName} was disqualified!"
o	You also have to reduce the attacker’s energy by 1. If it reaches 0, he/she is disqualified, and you need to remove him/her from your records and print the following message:
	"{attackerName} was disqualified!"
•	"Delete:{username}":
o	Delete all records of the given user, if he exists. If "All" is given as username - delete all records you have. 
In the end, you have to print the count of people left, each person with his/her health and energy sorted in descending order by the health and then by their name in ascending order in the following format:  
  People count: {count}
  {personName} - {health} - {energy}
  {personName} - {health} - {energy}